17 policies

+++ app/Http/Controllers/ArticleController.php
@@ -13,14 +13,16 @@ class ArticleController extends Controller
 {
     private static $messages = [
         'required' => 'El campo :attribute es obligatorio.',
-        'body.required' => 'Body no valido'
+        'body.required' => 'El Body no es valido'
     ];
     public function index()
     {
-        return new ArticleCollection(Article::paginate(25));
+        //$this->authorize('viewAny', Article::class);
+        return new ArticleCollection(Article::paginate(10));
     }
     public function show(Article $article)
     {
+        $this->authorize('view', $article);
         return response()->json(new ArticleResource($article),200);
     }
     public function image(Article $article)
@@ -29,6 +31,7 @@ class ArticleController extends Controller
     }
     public function store(Request $request)
     {
+        $this->authorize('create', Article::class);
         $request->validate([
         'title' => 'required|string|unique:articles|max:255',
         'body' => 'required',
@@ -58,6 +61,7 @@ class ArticleController extends Controller
     }
     public function update(Request $request,  Article $article)
     {
+        $this->authorize('update',$article);
         $request->validate([
             'title' => 'required|string|unique:articles,title,'.$article->id.'|max:255',
             'body' => 'required',
@@ -67,6 +71,7 @@ class ArticleController extends Controller
     }
     public function delete(Request $request,Article $article)
     {
+        $this->authorize('delete',$article);
         $article->delete();
         return response()->json(null, 204);
     }
